+++ client/src/game/ui/initiative/Initiative.vue
@@ -19,6 +19,7 @@ import {
     type InitiativeData,
     type InitiativeEffect,
     InitiativeEffectMode,
+    InitiativeEffectUpdateTiming,
     InitiativeSort,
     InitiativeTurnDirection,
 } from "../../models/initiative";
@@ -224,6 +225,15 @@ function createEffect(shape: GlobalId, effect: InitiativeEffect): void {
 
 function removeEffect(shape: GlobalId, index: number): void {
     if (initiativeStore.owns(shape)) initiativeStore.removeEffect(shape, index, true);
+}
+
+function invertEffectTiming(timing: InitiativeEffectUpdateTiming): InitiativeEffectUpdateTiming {
+    if (timing === InitiativeEffectUpdateTiming.TurnStart) return InitiativeEffectUpdateTiming.TurnEnd;
+    return InitiativeEffectUpdateTiming.TurnStart;
+}
+
+function changeEffectTiming(shape: GlobalId, index: number, timing: InitiativeEffectUpdateTiming): void {
+    if (initiativeStore.owns(shape)) initiativeStore.setEffectUpdateTiming(shape, index, timing, true);
 }
 
 function toggleGroupHighlight(actor: InitiativeData, show: boolean): void {
@@ -547,6 +557,42 @@ function n(e: any): number {
                                                         :key="`${actor.globalId}-${e}`"
                                                         class="initiative-effect-info"
                                                     >
+                                                        <div
+                                                            class="effect-icon-button"
+                                                            :class="{ disabled: !owns(actor.globalId) }"
+                                                            :title="
+                                                                t(
+                                                                    'game.ui.initiative.' +
+                                                                        (effect.updateTiming ===
+                                                                        InitiativeEffectUpdateTiming.TurnStart
+                                                                            ? 'start'
+                                                                            : 'end') +
+                                                                        '_turn_toggle_hint',
+                                                                )
+                                                            "
+                                                            @click="
+                                                                changeEffectTiming(
+                                                                    actor.globalId,
+                                                                    n(e),
+                                                                    invertEffectTiming(effect.updateTiming),
+                                                                )
+                                                            "
+                                                        >
+                                                            <font-awesome-icon
+                                                                icon="step-forward"
+                                                                :style="{
+                                                                    transform:
+                                                                        effect.updateTiming ===
+                                                                        InitiativeEffectUpdateTiming.TurnStart
+                                                                            ? 'scale(-1, 1)'
+                                                                            : 'scale(1, 1)',
+                                                                    cursor: !owns(actor.globalId)
+                                                                        ? 'default'
+                                                                        : 'pointer',
+                                                                }"
+                                                                style="opacity: 0.6"
+                                                            />
+                                                        </div>
                                                         <ResizingTextArea
                                                             v-model="effect.name"
                                                             :disabled="!owns(actor.globalId)"
@@ -574,8 +620,7 @@ function n(e: any): number {
                                                         </div>
                                                         <div
                                                             v-if="owns(actor.globalId)"
-                                                            class="actor-icon-button"
-                                                            style="margin-right: 4px"
+                                                            class="effect-icon-button"
                                                             :title="t('game.ui.initiative.delete_effect')"
                                                             @click="removeEffect(actor.globalId, n(e))"
                                                         >
@@ -969,6 +1014,7 @@ function n(e: any): number {
     }
 }
 
+.effect-icon-button,
 .actor-icon-button {
     display: flex;
     align-items: center;
@@ -995,6 +1041,11 @@ function n(e: any): number {
     }
 }
 
+.effect-icon-button {
+    padding: 2px 0;
+    font-size: 11pt;
+}
+
 .blurred {
     filter: blur(5px);
 }
@@ -1003,7 +1054,6 @@ function n(e: any): number {
     position: relative;
     display: flex;
     flex-direction: column;
-    width: -moz-fit-content;
     width: 15em;
     margin-right: 5px;
     border-bottom-left-radius: 5px;
@@ -1040,9 +1090,8 @@ function n(e: any): number {
 .initiative-effect-info {
     display: flex;
     flex-direction: row;
-    justify-content: flex-end;
+    justify-content: space-between;
     align-items: center;
-    padding-left: 5px;
     position: relative;
     z-index: 1;
     &::before {
@@ -1060,23 +1109,29 @@ function n(e: any): number {
     }
 
     > * {
+        margin: 0 3px;
+        &:first-child {
+            margin-left: 2px;
+        }
+        &:last-child {
+            margin-right: 2px;
+        }
+    }
+    > input {
         border: none;
         background-color: inherit;
         text-align: right;
         margin: 0 3px;
 
-        &:last-child {
-            margin-right: 0;
-        }
-    }
-    > input {
         font-size: 11pt;
     }
     .effect-turn-counter {
-        width: 25px;
+        min-width: 25px;
+        max-width: 25px;
         padding: 0 2px;
     }
     .infinite-placeholder {
+        text-align: right;
         user-select: none;
         font-size: 12pt;
     }