+++ server/src/api/models/initiative/effect.py
@@ -1,12 +1,20 @@
+from enum import IntEnum
+
 from pydantic import BaseModel, Field
 
 from ..helpers import TypeIdModel
+
+
+class InitiativeEffectUpdateTiming(IntEnum):
+    TurnEnd = 0
+    TurnStart = 1
 
 
 class ApiInitiativeEffect(BaseModel):
     name: str
     turns: str | None
     highlightsActor: bool
+    updateTiming: InitiativeEffectUpdateTiming
 
 
 class InitiativeEffectNew(TypeIdModel):
@@ -19,6 +27,12 @@ class InitiativeEffectRename(TypeIdModel):
     shape: str = Field(json_schema_extra={"typeId": "GlobalId"})
     index: int
     name: str
+
+
+class InitiativeEffectTiming(TypeIdModel):
+    shape: str = Field(json_schema_extra={"typeId": "GlobalId"})
+    index: int
+    timing: InitiativeEffectUpdateTiming
 
 
 class InitiativeEffectTurns(TypeIdModel):